(PC-24993)[PRO] refactor: remove closable style from base tag
