+++ pro/src/components/CollectiveOfferLayout/CollectiveOfferLayout.tsx
@@ -5,7 +5,8 @@ import { useLocation, useParams } from 'react-router-dom'
 import CollectiveOfferBreadcrumb from 'components/CollectiveOfferBreadcrumb'
 import HelpLink from 'components/HelpLink'
 import { getActiveStep } from 'pages/CollectiveOfferRoutes/utils/getActiveStep'
-import { Tag, Title } from 'ui-kit'
+import { Title } from 'ui-kit'
+import { Tag } from 'ui-kit/Tag/Tag'
 
 import styles from './CollectiveOfferLayout.module.scss'
