+++ pro/src/pages/Offers/Offers/OfferItem/Cells/OfferNameCell/OfferNameCell.tsx
@@ -8,8 +8,8 @@ import {
 import { Offer } from 'core/Offers/types'
 import { Audience } from 'core/shared'
 import fullErrorIcon from 'icons/full-error.svg'
-import { Tag } from 'ui-kit'
 import { SvgIcon } from 'ui-kit/SvgIcon/SvgIcon'
+import { Tag } from 'ui-kit/Tag/Tag'
 import { FORMAT_DD_MM_YYYY_HH_mm } from 'utils/date'
 import { pluralize } from 'utils/pluralize'
 import { formatLocalTimeDateString } from 'utils/timezone'
