+++ pro/src/ui-kit/Tag/Tag.module.scss
@@ -5,10 +5,6 @@
 .tag {
   @include fonts.highlight;
 
-  &:hover {
-    background-color: colors.$grey-medium;
-  }
-
   white-space: nowrap;
   justify-content: center;
   border: solid rem.torem(1px) colors.$tertiary;
@@ -17,34 +13,4 @@
   display: inline-flex;
   line-height: rem.torem(16px);
   padding: 0 rem.torem(4px) rem.torem(1px) rem.torem(4px);
-
-  &-closeable {
-    background-color: colors.$grey-light;
-    border-color: colors.$grey-light;
-    color: colors.$black;
-  }
-
-  &-close-button {
-    @include fonts.highlight;
-
-    line-height: rem.torem(16px);
-    background: none;
-    border: none;
-    display: flex;
-
-    &--disabled {
-      cursor: default;
-
-      .tag-close-button-icon {
-        cursor: default;
-      }
-    }
-
-    &-icon {
-      margin: rem.torem(3.2px) 0 rem.torem(1.6px) rem.torem(4.8px);
-      width: rem.torem(11.2px);
-      height: rem.torem(11.2px);
-      cursor: pointer;
-    }
-  }
 }
