+++ pro/src/ui-kit/Tag/Tag.stories.tsx
@@ -1,6 +1,6 @@
 import { StoryObj } from '@storybook/react'
 
-import Tag from './Tag'
+import { Tag } from './Tag'
 
 export default {
   title: 'ui-kit/Tag',
@@ -8,12 +8,6 @@ export default {
 }
 
 export const Default: StoryObj<typeof Tag> = {
-  args: {
-    children: 'Offre collective',
-  },
-}
-
-export const Closeable: StoryObj<typeof Tag> = {
   args: {
     children: 'Offre collective',
   },
