+++ pro/src/ui-kit/Tag/Tag.tsx
@@ -8,8 +8,6 @@ interface TagProps {
   children: ReactNode
 }
 
-const Tag = ({ children, className }: TagProps): JSX.Element => {
+export const Tag = ({ children, className }: TagProps): JSX.Element => {
   return <span className={cx(styles['tag'], className)}>{children}</span>
 }
-
-export default Tag
