--- pro/src/ui-kit/Tag/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Tag'
